refactor: Move stream/api/receive/jsonnet to stream/mapping/jsonnet
